delete rule tmp in Makefile
